+++ OneSignalExample/Assets/OneSignal/CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Removed `SetLaunchURLsInApp`
 - Removed async from location request permission and updated method name to `RequestPermission`
-- Updated included iOS SDK to [5.0.0-beta-06](https://github.com/OneSignal/OneSignal-iOS-SDK/releases/tag/5.0.0-beta-06)
+- Updated included iOS SDK to [5.0.1](https://github.com/OneSignal/OneSignal-iOS-SDK/releases/tag/5.0.1)
+- Updated included Android SDK to [5.0.0](https://github.com/OneSignal/OneSignal-Android-SDK/releases/tag/5.0.0)
 ### Fixed
 - Outcome method calls on Android