+++ OneSignalExample/Assets/OneSignal/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed async from location request permission and updated method name to `RequestPermission`
 - Updated included iOS SDK to [5.0.1](https://github.com/OneSignal/OneSignal-iOS-SDK/releases/tag/5.0.1)
 - Updated included Android SDK to [5.0.0](https://github.com/OneSignal/OneSignal-Android-SDK/releases/tag/5.0.0)
+- Updated default OneSignal Android notificaiton icons to new logo
 ### Fixed
 - Outcome method calls on Android