+++ OneSignalExample/Assets/OneSignal/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Sending VSAttribution data from the editor
+
 ## [5.0.2]
 ### Fixed
 - Stop foreground notifications from displaying after calling prevent default on iOS